% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipTo.R
\name{clipToExtent}
\alias{clipToExtent}
\title{clipToExtent}
\usage{
clipToExtent(sp, bb)
}
\arguments{
\item{sp}{an object of class "Spatial"}

\item{bb}{an object of class "matrix"}
}
\value{
an object clipped by the extent specified
}
\description{
Clips a spatial object given a geographic extent
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
