% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipTo.R
\name{clipToGlobalExtent}
\alias{clipToGlobalExtent}
\title{clipToGlobalExtent}
\usage{
clipToGlobalExtent(sp)
}
\arguments{
\item{sp}{an object of class "Spatial"}
}
\value{
an object clipped by the extent specified
}
\description{
Clips a spatial object given the global lon/lat extent
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
