% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCWPGrid.R
\name{createCWPGrid}
\alias{createCWPGrid}
\title{createCWPGrid}
\usage{
createCWPGrid(size = NULL, res = NULL, parallel = FALSE, ...)
}
\arguments{
\item{size}{an integer code corresponding to the grid size (referred as code
A in the CWP Handbook)}

\item{res}{a string matching one of the accepted resolution values. Accepted 
resolutions values are '10min_x_10min', '20min_x_20min', '30min_x_30min',
'30min_x_1deg', '1deg_x_1deg', '5deg_x_5deg', '10deg_x_10deg', '20deg_x_20deg',
'30deg_x_30deg'"}

\item{parallel}{run in parallel}

\item{...}{parallel options}
}
\value{
an object of class "SpatialPolygonsDataFrame"
}
\description{
Creates a CWP grid spatial object
}
\references{
CWP Handbook - http://www.fao.org/fishery/cwp/handbook/G/en
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
