% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportFeatures.R
\name{exportFeatures}
\alias{exportFeatures}
\title{exportFeatures}
\usage{
exportFeatures(features, outputFormat = "SHP", tozip = FALSE,
  file.path = NULL, file.name = NULL)
}
\arguments{
\item{features}{the sp object to export}

\item{outputFormat}{the format of the output file, by default "SHP"}

\item{tozip}{object of class "logical" indicating if a zip should be created.
Default is FALSE.}

\item{file.path}{the base path where to export the features}

\item{file.name}{the name of the output file}
}
\description{
Export sp object to well-known formats
}
\note{
only supported for GML and SHP
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
        Norbert Billet \email{norbert.billet@ird.fr}
}
