% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findP4s.r
\name{findP4s}
\alias{findP4s}
\title{findP4s}
\usage{
findP4s(srsName, morphToESRI = FALSE)
}
\arguments{
\item{srsName}{name of the spatial reference system}

\item{morphToESRI}{use the ESRI WKT representation}
}
\value{
a list of proj4 defnitions where the WKT GEOGCS attribute is equal to the required srsName
}
\description{
Find a proj4 definition from a srsName
}
\note{
We actually except lat/long coordinates systems due to confusion in output of different WFS vendors
}
\author{
Norbert Billet \email{norbert.billet@ird.fr}
        Emmanuel Blondel \email{emmanuel.blonde1@gmail.com}
}

