% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intersection.R
\name{intersection}
\alias{intersection}
\title{intersection}
\usage{
intersection(
  features1,
  features2,
  gmlIdAttributeName = c("gml_id", "gml_id"),
  areaCRS = NULL
)
}
\arguments{
\item{features1}{a first sp object}

\item{features2}{a second sp object}

\item{gmlIdAttributeName}{specific to GML, names of the ID attributes for each 
features, by default c("gml_id", "gml_id")}

\item{areaCRS}{a CRS object used as reference for area calculation. If no value 
is provided, take the current CRS. If NA is provided there no area computation.}
}
\value{
an object of class "Spatial"
}
\description{
Computes an Intersection
}
\note{
only supported for GML 2
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
        Norbert Billet \email{norbert.billet@ird.fr}
}
