% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/joinByArea.R
\name{joinByArea}
\alias{joinByArea}
\title{spatial join by area}
\usage{
joinByArea(x, features, refCode, filter = NULL, na.rm = TRUE)
}
\arguments{
\item{x}{an object of class "Spatial"}

\item{features}{an object of class "Spatial" from which we want to inherit a geocode}

\item{refCode}{an object of class = "character" representing the column to enrich x with}

\item{filter}{an optional filter in the form list(key="key",value="value") to apply to the 'features' object}

\item{na.rm}{TRUE (by default) if NA values (no spatial join) have to be removed}
}
\value{
an object of class "Spatial" (x enriched with the refCode from features)
}
\description{
performs a spatial join by area
}
\details{

}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}

