% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWFS.R
\name{readWFS}
\alias{readWFS}
\title{readWFS}
\usage{
readWFS(
  url,
  outputFormat = "GML",
  p4s = NULL,
  gmlIdAttributeName = "gml_id",
  target.dir = NULL,
  verbose = TRUE
)
}
\arguments{
\item{url}{the baseURL of the WFS GetFeature request}

\item{outputFormat}{the output format for the WFS GetFeature request, by default "GML"}

\item{p4s}{an optional proj4string, by default NULL (an attempt will be performed to get the projection from the data)}

\item{gmlIdAttributeName}{specific to GML, the name of the ID attribute, by default "gml_id"}

\item{target.dir}{a target directory where temporary GML files will be downloaded}

\item{verbose}{if log has to printed in the R console. TRUE by default}
}
\value{
an object of class "Spatial"
}
\description{
Read WFS & returns a sp object
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
        Norbert Billet \email{norbert.billet@ird.fr}
}
