% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reallocate.R
\name{reallocate}
\alias{reallocate}
\title{reallocate}
\usage{
reallocate(x, y, area.x, area.y, by.x = NULL, by.y = NULL, data, warea,
  wprob = NULL, aggregates = NULL)
}
\arguments{
\item{x}{a data.frame giving statistics reported for a given area}

\item{y}{a data.frame giving intersections between 2 or more types of areas, the intersecting surface
and possibly a giving a probability field}

\item{area.x}{name of the x-field representing the source area (typically an area code)}

\item{area.y}{name of the y-field representing the target area (typically an area code)}

\item{by.x}{name(s) of optional additional x-fields to be mapped}

\item{by.y}{name(s) of optional additional y-fields to be mapped}

\item{data}{name of the x-field representing the numerical values to reallocate}

\item{warea}{name of the y-field representing the intersecting surface}

\item{wprob}{name of the y-field representing an additional probability value. NULL by default}

\item{aggregates}{name(s) of the y-field(s) that represent to target geographic dimension}
}
\value{
an object of class "data.frame"
}
\description{
Computes a spatial reallocation of statistics
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}

