% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFigisGeo.R
\docType{package}
\name{RFigisGeo}
\alias{RFigisGeo}
\alias{RFigisGeo-package}
\title{Toos for geospatial data processing}
\description{
RFigisGeo provides a set of utility tools to do geospatial data processing
}
\details{
\tabular{ll}{
   Package: \tab RFigisGeo\cr
   Type: \tab Package\cr
   Version: \tab 0.3\cr
   Date: \tab 2018-08-15\cr
   License: \tab GPL(>=2.0)\cr
   LazyLoad: \tab yes\cr
 }
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
        Norbert Billet \email{norbert.billet@ird.fr}
}
