% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/toSpatialPoints.R
\name{toSpatialPoints}
\alias{toSpatialPoints}
\title{toSpatialPoints}
\usage{
toSpatialPoints(x, lonlat, verbose = TRUE)
}
\arguments{
\item{x}{an object of class "data.frame"}

\item{lonlat}{a vector of length 2 given the lon/lat column names e.g. c("Lon","Lat")}

\item{verbose}{TRUE (by default) to display logs}
}
\value{
an object of class "SpatialPoints"
}
\description{
Convert a data.frame with lonlat columns to a SpatialPoints object
}
\details{

}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}

