% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exportFeatures.R
\name{exportFeatures}
\alias{exportFeatures}
\title{exportFeatures}
\usage{
exportFeatures(features, outputFormat = "SHAPE-ZIP", file.path = NULL,
  file.name = NULL)
}
\arguments{
\item{features}{the sp object to export}

\item{outputFormat}{the format of the output file, by default "SHAPE-ZIP" (a zipped shapefile)}

\item{file.path}{the base path where to export the features}

\item{file.name}{the name of the output file}

\item{only}{supported for GML and SHAPE-ZIP}
}
\description{
Export sp object to well-known formats
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
        Norbert Billet \email{norbert.billet@ird.fr}
}

